requirements.txt
pandas
IPython
tensorflow==2.8.0
opencv-python
pillow
pyyaml
scikit-image
tqdm
scikit-learn
plotly
matplotlib